wibestore_backend/requirements.txt
# Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
drf-spectacular>=0.27,<1.0

# Database
psycopg[binary]>=3.1,<4.0

# Cache & Message Broker
redis>=5.0,<6.0

# Async Tasks
celery>=5.3,<6.0
django-celery-beat>=2.6,<3.0

# WebSocket
channels>=4.2,<5.0
channels-redis>=4.2,<5.0

# Media
Pillow>=10.0

# Storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Payments
stripe>=8.0,<9.0

# Servers
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.29,<1.0

# Environment
django-environ>=0.11,<1.0

# Security
cryptography>=42.0,<43.0
django-axes>=6.4,<7.0
argon2-cffi>=23.1,<24.0

# OAuth
social-auth-app-django>=5.4,<6.0

# Monitoring
sentry-sdk[django]>=1.40,<2.0
django-health-check>=3.18,<4.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
pytest-asyncio>=0.23,<1.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0
coverage>=7.4,<8.0
black>=24.0,<25.0
ruff>=0.2,<1.0
isort>=5.13,<6.0
pre-commit>=3.6,<4.0
django-debug-toolbar>=4.3,<5.0
bandit>=1.7,<2.0
whitenoise>=6.6.0
